app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
opencv-python
pytesseract
Pillow
